Moved fixtures and consts
